requirements.txt
# Core data science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Time series models
statsmodels>=0.14.0
prophet>=1.1.5
xgboost>=2.0.0
tensorflow>=2.13.0   # CPU build is fine

# Feature engineering
holidays>=0.40
python-dateutil>=2.8.2

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0

# Persistence
joblib>=1.3.0

# Visualisation
matplotlib>=3.7.0
seaborn>=0.13.0
plotly>=5.18.0

# Testing
pytest>=7.4.0
httpx>=0.25.0       # for FastAPI test client
